+++ src/viser/client/src/App.tsx
@@ -72,7 +72,7 @@ export type ViewerContextContents = {
     [name: string]:
       | undefined
       | {
-          poseUpdateState?: "updated" | "needsUpdate" | "waitForMakeObject";
+          poseUpdateState?: "updated" | "needsUpdate";
           wxyz?: [number, number, number, number];
           position?: [number, number, number];
           visibility?: boolean; // Visibility state from the server.
@@ -470,7 +470,7 @@ function AdaptiveDpr() {
       iterations={5}
       step={0.1}
       bounds={(refreshrate) => {
-        const max = Math.min(refreshrate * 0.9, 85);
+        const max = Math.min(refreshrate * 0.75, 85);
         const min = Math.max(max * 0.5, 38);
         return [min, max];
       }}
